Initial commit
A program that cuts and adds to a string from back to front based on user input provided by two users. The program also keeps track of the iterations it takes to finish the task based on the provided inputs. Note: the program should be terminated after each use as it will continue to count on top of the old iterations if the memory catch is not cleared.

The program still needs more work mainly with the control of input and code complexity.